fix(main, common): minor fixes, moved back args.go and env_store.go to main
